+++ release_notes.md
@@ -47,6 +47,7 @@
 
 ## Features
 1. [TH2-4869] Added ResponseFormat class to provide possible values for response_formats commands parameter. 
+2. [TH2-4692] Added response_formats for GetMessageById and GetMessagesById
 
 ## Improvements
 1. [TH2-4869] ResponseFormat == 'JSON_PARSED' made as default